+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased
 -   Drop support for Python 2, 3.4, and 3.5.
 -   Bump minimum version of Flask to 1.0.4.
 -   Bump minimum version of SQLAlchemy to 1.2.
+-   Add mypy to the build and add typing for some components
 -   Remove previously deprecated code.
 -   The CamelCase to snake_case table name converter handles more
     patterns correctly. If such a name was already created in the